requirements.txt

Flask==3.0.0
flask-restx==1.3.0
SQLAlchemy>=2.0.35
python-dotenv==1.0.0





pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
isort==5.13.2


colorlog==6.8.0
